[refactor] src 폴더 구조 변경 - apis, styles 폴더 구조 변경
