--- src/pages/styles.ts
@@ -1,17 +1,0 @@
-import styled from '@emotion/styled';
-
-const Container = styled.div`
-  width: 80%;
-  max-width: 1200px;
-  margin: 0 auto;
-`;
-
-const Divider = styled.hr`
-  width: 100%;
-  position: absolute;
-  left: 0;
-  color: rgba(0, 0, 0, 0.15);
-  margin: 0;
-`;
-
-export { Container, Divider };
